Create login page and dashboard page. Redirect users to dashboard if logged in and to login page if not.
