+++ src/app/layout.tsx
@@ -1,7 +1,5 @@
 import "~/styles/globals.css";
-
 import { Inter } from "next/font/google";
-
 import { TRPCReactProvider } from "~/trpc/react";
 
 const inter = Inter({
